requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
Cython==0.29.34
fastapi==0.95.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
joblib==1.2.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
patsy==0.5.3
pmdarima==2.0.3
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
statsmodels==0.14.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.22.0
